Add prettier plugin to solhint
- Ensures consistent styling

+++ contracts/interfaces/ICERC20.sol
@@ -2,9 +2,17 @@ pragma solidity 0.6.12;
 
 interface ICERC20 {
     function mint(uint256) external returns (uint256);
+
     function exchangeRateCurrent() external returns (uint256);
+
     function supplyRatePerBlock() external returns (uint256);
-    function redeem(uint) external returns (uint);
-    function redeemUnderlying(uint) external returns (uint);
-    function balanceOfUnderlying(address account) external view returns (uint);
+
+    function redeem(uint256) external returns (uint256);
+
+    function redeemUnderlying(uint256) external returns (uint256);
+
+    function balanceOfUnderlying(address account)
+        external
+        view
+        returns (uint256);
 }
